requirements.txt
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
pyserial>=3.5
pyautogui>=0.9.53
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.1.0
tqdm>=4.62.0
python-dotenv>=0.19.0
tkinter
pytest>=7.0.0
pytest-cov>=3.0.0